pyproject.toml
[project]
name = "multimodal"
version = "0.0.1"

description = "School project"

authors = [
  { name = "ewkt" },
]

readme = "README.md"
requires-python = "~=3.10"

dependencies = [
    "pip==25.0",
    "python-dotenv",
    "ruff==0.11.0",
    "tqdm",
    "typer==0.15.1",
    "transformers==4.48.3",
    "sentence-transformers==3.4.0",
    "faiss-cpu==1.9.0",
    "PyPDF2==2.10.3",
    "pillow==11.1.0",
    "torch==2.5.1",
]


classifiers = [
    "Programming Language :: Python :: 3",
]

[dependency-groups]
test = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.23.5",
]